refactoring: 퀘스트 생성 기능 hook 분리
