+++ client/src/pages/CreateMainQuest.tsx
@@ -6,38 +6,29 @@ import { FiMinusCircle } from 'react-icons/fi';
 import Button from '@/components/common/Button';
 import QuestInputBox from '@/components/quests/QuestInputBox';
 import { media } from '@/styles/theme';
-import { useState } from 'react';
 import CloseButton from '@/components/common/CloseButton';
-import { CreateMainQuestProps, useMainQuest } from '@/hooks/useMainQuest';
-import { QuestDifficulty, QuestHiddenType, QuestMode } from '@/models/quest.model';
-import { useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
+import { useCreateMainQuestForm } from '@/hooks/useMainQuest';
 
 const CreateMainQuest = () => {
-  const [isPrivate, setIsPrivate] = useState(false);
-  const [isDifficulty, setIsDifficulty] = useState(0);
-  const [startDate, setStartDate] = useState('');
-  const [endDate, setEndDate] = useState('');
-  const [plusQuest, setPlusQuest] = useState(1);
-  const [minusQuest, setMinusQuest] = useState(0);
-  const today = new Date().toISOString().substring(0, 10);
+  const {
+    register,
+    isPrivate,
+    setIsPrivate,
+    isDifficulty,
+    setIsDifficulty,
+    plusQuest,
+    setPlusQuest,
+    minusQuest,
+    setMinusQuest,
+    startDate,
+    setStartDate,
+    endDate,
+    setEndDate,
+    today,
+    handleSubmit,
+  } = useCreateMainQuestForm();
   const navigate = useNavigate();
-  const { CreateQuestMutation } = useMainQuest();
-
-  const { register, control, handleSubmit } = useForm<CreateMainQuestProps>();
-
-  const onSubmit = handleSubmit((data) => {
-    if (data.sideQuests && data.sideQuests.length > 0) {
-      const hidden = (isPrivate ? 'TRUE' : 'FALSE') as QuestHiddenType;
-      const difficulty =
-        isDifficulty === 0 ? 'EASY' : isDifficulty === 1 ? 'NORMAL' : ('HARD' as QuestDifficulty);
-      const mode = 'MAIN' as QuestMode;
-      const newData = { ...data, hidden, difficulty: difficulty, mode: mode };
-      CreateQuestMutation.mutate(newData);
-    } else {
-      alert('사이드 퀘스트를 추가해주세요.');
-    }
-  });
 
   return (
     <>
@@ -53,7 +44,7 @@ const CreateMainQuest = () => {
             )}
           </div>
         </header>
-        <form onSubmit={onSubmit}>
+        <form onSubmit={handleSubmit}>
           <QuestInputBox placeholder="퀘스트 제목" {...register('title')} />
           <QuestButtonContainer>
             <Button
